requirements.txt
# ============================================
# Podcast Factory - 依赖清单
# ============================================

# ===== API & Web 框架 =====
fastapi==0.109.2
uvicorn[standard]==0.27.1

# ===== LLM & 编排 =====
langgraph==0.1.2
openai>=1.0.0

# ===== HTTP 客户端 =====
httpx==0.26.0

# ===== 数据验证 & 配置 =====
pydantic==2.6.1
python-dotenv==1.0.1

# ===== 定时任务 =====
apscheduler==3.10.4

# ===== 音频处理 =====
pydub==0.25.1

# ===== 日志 =====
loguru==0.7.2

# ===== 开发 & 测试 =====
pytest>=8.0.0
pytest-asyncio>=0.24.0
